# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/subscriptions:
    get:
      operationId: listStudentSubscriptions
      tags: [student-subscriptions]
      summary: List own subscriptions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentSubscription"
    post:
      operationId: createStudentSubscription
      tags: [student-subscriptions]
      summary: Purchase a subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSubscriptionResponse"
        "404":
          description: Restaurant or package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /student/subscriptions/{subscriptionId}:
    get:
      operationId: getStudentSubscription
      tags: [student-subscriptions]
      summary: Get a subscription by ID
      security:
        - bearerAuth: []
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentSubscription"
  /student/subscriptions/{subscriptionId}/cancel:
    post:
      operationId: cancelStudentSubscription
      tags: [student-subscriptions]
      summary: Cancel a subscription
      security:
        - bearerAuth: []
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  refundAmount:
                    type: number
                required: [message, refundAmount]
  /student/subscriptions/{subscriptionId}/pause:
    post:
      operationId: pauseStudentSubscription
      tags: [student-subscriptions]
      summary: Pause a subscription for N days
      security:
        - bearerAuth: []
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pauseDays:
                  type: integer
                  minimum: 1
                  maximum: 30
              required: [pauseDays]
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  pausedUntil:
                    type: string
                required: [message, pausedUntil]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    StudentSubscription:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        restaurantName:
          type: string
        packageId:
          type: string
          nullable: true
        packageName:
          type: string
        slot:
          type: string
          enum: [lunch, dinner, both]
        totalDays:
          type: integer
        usedDays:
          type: integer
        remainingDays:
          type: integer
        pricePerDay:
          type: number
        totalPaid:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, paused, cancelled, completed, refund_requested]
        pausedUntil:
          type: string
          nullable: true
        lateCancellationFees:
          type: number
        createdAt:
          type: string
      required: [id, restaurantId, restaurantName, packageName, slot, totalDays, usedDays, remainingDays, pricePerDay, totalPaid, startDate, endDate, status, lateCancellationFees, createdAt]
    CreateSubscriptionBody:
      type: object
      properties:
        restaurantId:
          type: string
        packageId:
          type: string
        startDate:
          type: string
          description: "YYYY-MM-DD, defaults to today"
      required: [restaurantId, packageId]
    CreateSubscriptionResponse:
      type: object
      properties:
        subscription:
          $ref: "#/components/schemas/StudentSubscription"
        ordersCreated:
          type: integer
        message:
          type: string
      required: [subscription, ordersCreated, message]
...